+++ eval-fox-eval/src/main/java/com/evalkit/framework/eval/facade/EvalFacade.java
@@ -1,9 +1,62 @@
 package com.evalkit.framework.eval.facade;
 
 /**
- * 评测门面
+ * 抽象评测门面
+ * 评测生命周期: 初始化 -> 执行前 -> 执行(加载数据前 -> 加载数据 -> 加载数据后 -> 评测前 -> 评测 -> 评测后 -> 结果上报前 -> 结果上报 -> 结果上报后) -> 执行后
  */
 public abstract class EvalFacade {
+
+    /**
+     * 评测入口
+     */
+    public void run() {
+        init();
+        executeWrapper();
+    }
+
+    /**
+     * 初始化
+     */
+    protected abstract void init();
+
+    /**
+     * 评测执行生命周期
+     */
+    public void executeWrapper() {
+        beforeExecute();
+        execute();
+        afterExecute();
+    }
+
+    /**
+     * 执行前钩子
+     */
+    protected void beforeExecute() {
+    }
+
+    /**
+     * 评测执行周期
+     */
+    protected void execute() {
+        loadDataWrapper();
+        evalWrapper();
+        reportWrapper();
+    }
+
+    /**
+     * 执行后钩子
+     */
+    protected void afterExecute() {
+    }
+
+    /**
+     * 加载数据生命周期
+     */
+    protected void loadDataWrapper() {
+        beforeLoadData();
+        loadData();
+        afterLoadData();
+    }
 
     /**
      * 加载数据前钩子
@@ -23,12 +76,12 @@ public abstract class EvalFacade {
     protected abstract void loadData();
 
     /**
-     * 加载数据生命周期
+     * 评测生命周期
      */
-    protected void loadDataWrapper() {
-        beforeLoadData();
-        loadData();
-        afterLoadData();
+    protected void evalWrapper() {
+        beforeEval();
+        eval();
+        afterEval();
     }
 
     /**
@@ -47,15 +100,6 @@ public abstract class EvalFacade {
      * 评测
      */
     protected abstract Object eval();
-
-    /**
-     * 评测生命周期
-     */
-    protected void evalWrapper() {
-        beforeEval();
-        eval();
-        afterEval();
-    }
 
     /**
      * 结果上报前钩子
@@ -83,36 +127,6 @@ public abstract class EvalFacade {
         beforeReport();
         report();
         afterReport();
-    }
-
-    /**
-     * 评测执行周期
-     */
-    public void execute() throws Exception {
-        beforeExecute();
-        doExecute();
-        afterExecute();
-    }
-
-    /**
-     * 实际操作
-     */
-    protected void doExecute() {
-        loadDataWrapper();
-        evalWrapper();
-        reportWrapper();
-    }
-
-    /**
-     * 执行前钩子
-     */
-    protected void beforeExecute() {
-    }
-
-    /**
-     * 执行后钩子
-     */
-    protected void afterExecute() {
     }
 
     /**