+++ docs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # ChangeLog
 
+## [0.1.1] - 2025-10-09
+
+### Added
+- 增加是否开启断点续评开关, 默认开启
+- 多数据加载器增加配置和分页构造方法
+### Fixed
+- 修复评测结果错误的bug
+- 使用fastjson2替换jayway.jsonpath解析json路径
+- 修复增量评测消息消费bug
+- 修复html结果上报的未关闭writer的bug
+- 修复文件上报时空指针bug
+- 优化评测门面执行生命周期
+### Changed
+- pom统一版本号管理
+- 添加flatten-maven-plugin插件,打包时自动展开变量
+
 ## [0.1.0] - 2025-10-08
 
 ### Added
@@ -11,6 +27,7 @@
 - 弃用ThreadLocal管理上下文,改为变量传递
 ### Changed
 - 删除无效单测
+- 
 
 ## [0.0.1] - 2025-09-27